model loader lib initial implementation
