+++ include/render/render.h
@@ -14,8 +14,6 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
-#include <tiny_obj_loader.h>
-
 #include <iostream>
 #include <fstream>
 #include <stdexcept>
@@ -29,7 +27,10 @@
 #include <set>
 #include <array>
 #include <unordered_map>
+
 #include <shadermanager/ShaderManager.h>
+#include <modelloader/modelloader.h>
+#include <modelloader/Vertex.h>
 
 const uint32_t WIDTH = 800;
 const uint32_t HEIGHT = 600;
@@ -38,6 +39,7 @@ const int MAX_FRAMES_IN_FLIGHT = 3;
 const std::string MODEL_PATH = "misc/cube.obj";
 const std::string TEXTURE_PATH = "misc/white.jpg";
 const std::string MTL_PATH = "misc/";
+nevk::Model testmodel;
 
 const std::vector<const char*> validationLayers = {
     "VK_LAYER_KHRONOS_validation"
@@ -92,80 +94,6 @@ struct SwapChainSupportDetails
     std::vector<VkSurfaceFormatKHR> formats;
     std::vector<VkPresentModeKHR> presentModes;
 };
-
-struct Vertex
-{
-    glm::vec3 pos;
-    glm::vec3 color;
-    glm::vec3 ka;
-    glm::vec3 kd;
-    glm::vec3 ks;
-    glm::vec2 texCoord;
-
-    static VkVertexInputBindingDescription getBindingDescription()
-    {
-        VkVertexInputBindingDescription bindingDescription{};
-        bindingDescription.binding = 0;
-        bindingDescription.stride = sizeof(Vertex);
-        bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-        return bindingDescription;
-    }
-
-    static std::array<VkVertexInputAttributeDescription, 5> getAttributeDescriptions()
-    {
-        std::array<VkVertexInputAttributeDescription, 5> attributeDescriptions{};
-
-        attributeDescriptions[0].binding = 0;
-        attributeDescriptions[0].location = 0;
-        attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[0].offset = offsetof(Vertex, pos);
-
-        /*attributeDescriptions[1].binding = 0;
-        attributeDescriptions[1].location = 1;
-        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[1].offset = offsetof(Vertex, color);*/
-
-        attributeDescriptions[1].binding = 0;
-        attributeDescriptions[1].location = 1;
-        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[1].offset = offsetof(Vertex, ka);
-
-        attributeDescriptions[2].binding = 0;
-        attributeDescriptions[2].location = 2;
-        attributeDescriptions[2].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[2].offset = offsetof(Vertex, kd);
-
-        attributeDescriptions[3].binding = 0;
-        attributeDescriptions[3].location = 3;
-        attributeDescriptions[3].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[3].offset = offsetof(Vertex, ks);
-
-        attributeDescriptions[4].binding = 0;
-        attributeDescriptions[4].location = 4;
-        attributeDescriptions[4].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[4].offset = offsetof(Vertex, texCoord);
-
-        return attributeDescriptions;
-    }
-
-    bool operator==(const Vertex& other) const
-    {
-        return pos == other.pos && ka == other.ka && kd == other.kd && ks == other.ks && texCoord == other.texCoord;
-    }
-};
-
-namespace std
-{
-template <>
-struct hash<Vertex>
-{
-    size_t operator()(Vertex const& vertex) const
-    {
-        return ((hash<glm::vec3>()(vertex.pos) ^ (hash<glm::vec3>()(vertex.color) << 1)) >> 1) ^ (hash<glm::vec2>()(vertex.texCoord) << 1);
-    }
-};
-} // namespace std
 
 struct UniformBufferObject
 {
@@ -292,7 +220,7 @@ private:
         createTextureImage();
         createTextureImageView();
         createTextureSampler();
-        loadModel();
+        loadModels();
         createVertexBuffer();
         createIndexBuffer();
         createUniformBuffers();
@@ -1151,77 +1079,11 @@ private:
         endSingleTimeCommands(commandBuffer);
     }
 
-
-    void loadModel()
+    void loadModels()
     {
-        tinyobj::attrib_t attrib;
-        std::vector<tinyobj::shape_t> shapes;
-        std::vector<tinyobj::material_t> materials;
-        std::string warn, err;
-
-        if (!tinyobj::LoadObj(&attrib, &shapes, &materials, &warn, &err, MODEL_PATH.c_str(), MTL_PATH.c_str()))
-        {
-            throw std::runtime_error(warn + err);
-        }
-
-        std::unordered_map<Vertex, uint32_t> uniqueVertices{};
-
-        //for (const auto& shape : shapes)
-        for (size_t s = 0; s < shapes.size(); s++)
-        {
-            size_t index_offset = 0;
-            // for (const auto& index : shape.mesh.indices)
-            for (size_t f = 0; f < shapes[s].mesh.num_face_vertices.size(); f++)
-            {
-                int fv = shapes[s].mesh.num_face_vertices[f];
-                for (size_t v = 0; v < fv; v++)
-                {
-                    Vertex vertex{};
-                    auto idx = shapes[s].mesh.indices[index_offset + v];
-                    vertex.pos = {
-                        attrib.vertices[3 * idx.vertex_index + 0],
-                        attrib.vertices[3 * idx.vertex_index + 1],
-                        attrib.vertices[3 * idx.vertex_index + 2]
-                    };
-
-                    vertex.texCoord = {
-                        attrib.texcoords[2 * idx.texcoord_index + 0],
-                        1.0f - attrib.texcoords[2 * idx.texcoord_index + 1]
-                    };
-
-                    vertex.color = { attrib.colors[3 * idx.vertex_index + 0],
-                                     attrib.colors[3 * idx.vertex_index + 1],
-                                     attrib.colors[3 * idx.vertex_index + 2] };
-
-                    vertex.ka = { materials[shapes[s].mesh.material_ids[f]].ambient[0],
-                                  materials[shapes[s].mesh.material_ids[f]].ambient[1],
-                                  materials[shapes[s].mesh.material_ids[f]].ambient[2]
-
-                    };
-
-                    vertex.kd = { materials[shapes[s].mesh.material_ids[f]].diffuse[0],
-                                  materials[shapes[s].mesh.material_ids[f]].diffuse[1],
-                                  materials[shapes[s].mesh.material_ids[f]].diffuse[2]
-
-                    };
-
-                    vertex.ks = { materials[shapes[s].mesh.material_ids[f]].specular[0],
-                                  materials[shapes[s].mesh.material_ids[f]].specular[1],
-                                  materials[shapes[s].mesh.material_ids[f]].specular[2]
-
-                    };
-
-                    if (uniqueVertices.count(vertex) == 0)
-                    {
-                        uniqueVertices[vertex] = static_cast<uint32_t>(vertices.size());
-                        vertices.push_back(vertex);
-                    }
-
-                    indices.push_back(uniqueVertices[vertex]);
-                }
-                index_offset += fv;
-            }
-        }
+        testmodel.loadModel(MODEL_PATH, MTL_PATH);
+        vertices = testmodel.getVertices();
+        indices = testmodel.getIndices();
     }
 
     void createVertexBuffer()
